requirements.txt
flask
gunicorn
pandas
yfinance
alpaca-py
pyarrow
websocket-client
numpy
scipy
matplotlib
seaborn
yfinance
cvxpy
Pyportfolioopt
qiskit
tensorflow
keras
torch
scikit-learn
jupyter
pandas
pandas-datareader
